trying to fix memory issues in tests
