+++ geomesa-accumulo/geomesa-accumulo-datastore/src/test/scala/org/locationtech/geomesa/accumulo/data/AccumuloDataStoreIdlTest.scala
@@ -1,0 +1,90 @@
+/*
+ * Copyright (c) 2013-2015 Commonwealth Computer Research, Inc.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Apache License, Version 2.0 which
+ * accompanies this distribution and is available at
+ * http://www.opensource.org/licenses/apache2.0.php.
+ */
+
+package org.locationtech.geomesa.accumulo.data
+
+import com.vividsolutions.jts.geom.Coordinate
+import org.geotools.data._
+import org.geotools.factory.{CommonFactoryFinder, Hints}
+import org.geotools.feature.simple.SimpleFeatureBuilder
+import org.geotools.feature.{DefaultFeatureCollection, NameImpl}
+import org.geotools.filter.text.cql2.CQL
+import org.geotools.filter.text.ecql.ECQL
+import org.geotools.geometry.jts.JTSFactoryFinder
+import org.geotools.referencing.CRS
+import org.geotools.referencing.crs.DefaultGeographicCRS
+import org.joda.time.{DateTime, DateTimeZone}
+import org.junit.runner.RunWith
+import org.locationtech.geomesa.accumulo.{TestWithDataStore, TestWithMultipleSfts}
+import org.locationtech.geomesa.accumulo.index.QueryHints._
+import org.locationtech.geomesa.accumulo.index.Strategy.StrategyType
+import org.locationtech.geomesa.accumulo.index.{ExplainString, JoinPlan, QueryPlanner}
+import org.locationtech.geomesa.accumulo.iterators.{BinAggregatingIterator, TestData}
+import org.locationtech.geomesa.features.ScalaSimpleFeature
+import org.locationtech.geomesa.filter.function.Convert2ViewerFunction
+import org.locationtech.geomesa.utils.filters.Filters
+import org.locationtech.geomesa.utils.geotools.Conversions._
+import org.locationtech.geomesa.utils.geotools.SimpleFeatureTypes
+import org.locationtech.geomesa.utils.text.WKTUtils
+import org.opengis.filter.Filter
+import org.specs2.mutable.Specification
+import org.specs2.runner.JUnitRunner
+import org.specs2.time.Duration
+
+import scala.collection.JavaConversions._
+import scala.util.Random
+
+@RunWith(classOf[JUnitRunner])
+class AccumuloDataStoreIdlTest extends Specification with TestWithDataStore {
+
+  sequential
+
+  override val spec = TestData.getTypeSpec()
+
+  addFeatures(TestData.allThePoints.map(TestData.createSF))
+
+  val ff = CommonFactoryFinder.getFilterFactory2
+  val srs = CRS.toSRS(DefaultGeographicCRS.WGS84)
+
+  "AccumuloDataStore" should {
+
+    "handle IDL correctly" in {
+      "default layer preview, bigger than earth, multiple IDL-wrapping geoserver BBOX" in {
+        val spatial = ff.bbox("geom", -230, -110, 230, 110, srs)
+        val query = new Query(sft.getTypeName, spatial)
+        val results = fs.getFeatures(query)
+        results.size() mustEqual 361
+      }
+
+      "greater than 180 lon diff non-IDL-wrapping geoserver BBOX" in {
+        val spatial = ff.bbox("geom", -100, 1.1, 100, 4.1, srs)
+        val query = new Query(sft.getTypeName, spatial)
+        val results = fs.getFeatures(query)
+        results.size() mustEqual 6
+      }
+
+      "small IDL-wrapping geoserver BBOXes" in {
+        val spatial1 = ff.bbox("geom", -181.1, -90, -175.1, 90, srs)
+        val spatial2 = ff.bbox("geom", 175.1, -90, 181.1, 90, srs)
+        val binarySpatial = ff.or(spatial1, spatial2)
+        val query = new Query(sft.getTypeName, binarySpatial)
+        val results = fs.getFeatures(query)
+        results.size() mustEqual 10
+      }
+
+      "large IDL-wrapping geoserver BBOXes" in {
+        val spatial1 = ff.bbox("geom", -181.1, -90, 40.1, 90, srs)
+        val spatial2 = ff.bbox("geom", 175.1, -90, 181.1, 90, srs)
+        val binarySpatial = ff.or(spatial1, spatial2)
+        val query = new Query(sft.getTypeName, binarySpatial)
+        val results = fs.getFeatures(query)
+        results.size() mustEqual 226
+      }
+    }
+  }
+}
